tests/token: Move tests into `routes` submodules
