+++ src/tests/routes/me/mod.rs
@@ -1,0 +1,1 @@
+pub mod tokens;
