+++ src/tests/routes/me/tokens/mod.rs
@@ -1,0 +1,4 @@
+pub mod create;
+pub mod delete;
+pub mod delete_current;
+pub mod list;
