+++ src/tests/routes/mod.rs
@@ -14,4 +14,5 @@
 pub mod categories;
 pub mod category_slugs;
 pub mod crates;
+pub mod me;
 pub mod versions;
